consolidate db.constraints module
