--- db/constraints/operations/select.py
@@ -1,5 +1,0 @@
-from db.connection import select_from_msar_func
-
-
-def get_constraints_for_table(table_oid, conn):
-    return select_from_msar_func(conn, 'get_constraints_for_table', table_oid)
